Spring Security với Json Web Token

+++ src/main/java/com/dong/jobhunter/controller/UserController.java
@@ -2,9 +2,12 @@ package com.dong.jobhunter.controller;
 
 import com.dong.jobhunter.domain.User;
 import com.dong.jobhunter.service.UserService;
+import com.dong.jobhunter.service.error.IdInvalidException;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,28 +21,40 @@ public class UserController {
     UserService userService;
 
     @GetMapping
-    public List<User> getUsers() {
-        return userService.getUser();
+    public ResponseEntity<List<User>> getUsers() {
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(userService.getUser());
     }
 
     @GetMapping("/{id}")
-    public User getUserById(@PathVariable Long id) {
-        return userService.getUserById(id.intValue());
+    public ResponseEntity<User> getUserById(@PathVariable Long id) {
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(userService.getUserById(id.intValue()));
     }
 
     @PostMapping("")
-    public User createNewUser(@RequestBody User userRequest) {
-        return userService.createNewUser(userRequest);
+    public ResponseEntity<User> createNewUser(@RequestBody User userRequest) {
+
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(userService.createNewUser(userRequest));
     }
 
     @PutMapping("/{id}")
-    public User updateUser(@PathVariable Long id, @RequestBody User userRequest) {
-        return userService.updateUser(id, userRequest);
+    public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User userRequest) {
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(userService.updateUser(id, userRequest));
     }
 
     @DeleteMapping("/{id}")
-    public String deleteUser(@PathVariable Long id) {
+    public ResponseEntity<String> deleteUser(@PathVariable Long id)
+            throws IdInvalidException {
+        if (id >= 1500) {
+            throw new IdInvalidException("Id khong lon hon 1500");
+        }
         userService.deleteUser(id);
-        return "Deleted Successfully";
+        return ResponseEntity.status(HttpStatus.NO_CONTENT)
+                .body("Deleted Successfully");
     }
 }
